requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
pydantic>=2.0.0
pyyaml>=6.0
scikit-image>=0.20.0
scipy>=1.10.0
matplotlib>=3.7.0
tqdm>=4.65.0
openpyxl>=3.1.0
pytest>=7.0.0
